Moved profile Menu and Card
