--- app/components/Menu/Item.tsx
@@ -1,5 +1,0 @@
-export const MenuItem = ({value}: {value: string}): JSX.Element => (
-  <div className='h-12 border-gray-400 border-2 p-2'>
-    <span>{value}</span>
-  </div>
-)
